requirements.txt
discord==1.0.1
discord.py==1.3.2
python-dotenv==0.12.0
psycopg2-binary==2.8.3
google-search-results-serpwow==1.1.8